fixed bugs related to rf

+++ prog/code/hydra_alpha_v1/src/tank/peripheral/protocol_format/IncommingCommunicationFormat.h
@@ -8,8 +8,8 @@
 typedef struct
 {
     uint8_t mode;
-    int8_t throttle;
-    int8_t steer;
+    float throttle;
+    float steer;
     float sonarAngleZ;
     float cameraAngleZ;
     float cameraAngleY;
